+++ src/Controller/TrickDeleteController.php
@@ -2,13 +2,10 @@
 
 namespace App\Controller;
 
-use App\Entity\Trick;
-use App\Repository\TrickRepository;
-use App\Repository\VideoRepository;
-use App\Repository\ImageRepository;
+use App\Entity\{Trick, Video, Image};
+use App\Repository\{TrickRepository, ImageRepository, VideoRepository};
 use Symfony\Component\HttpFoundation\Response;
 use App\Tool\DeleteFile;
-use App\Responders\TrickAddResponder;
 use Symfony\Component\Filesystem\Filesystem;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
@@ -17,6 +14,7 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class TrickDeleteController extends AbstractController
 {
@@ -27,63 +25,63 @@ class TrickDeleteController extends AbstractController
     /** @var EntityManagerInterface */
     private $session;
     private $deleteFile;
+    private $tokenStorage;
 
     public function __construct(
         SessionInterface $session,
-        DeleteFile $deleteFile
-    
+        DeleteFile $deleteFile,
+        TokenStorageInterface $tokenStorage
     ){
         $this->session = $session;
-        $this->deleteFile =$deleteFile;
+        $this->deleteFile = $deleteFile;
+        $this->tokenStorage = $tokenStorage;
     }
 
     /**
     * @Route("/delete_trick/{id}", name="delete_trick")
     * @IsGranted("ROLE_ADMIN")
     */
-    public function delete($id, TrickRepository $trickRepository, VideoRepository $videoRepository, ImageRepository $imageRepository)
+    public function delete($id)
     {
-        // creates a task object and initializes some data for this example
-        //$task->setTask('Write a blog post');
-        //$task->setDueDate(new \DateTime('tomorrow'));
-        //var_dump($id);
-        $video = $videoRepository
-        ->findOneBy(['trickId' => $id]);
-        $trick = $trickRepository
-        ->find($id);
-        $img = $imageRepository->findBy(['trickId' => $id]);
-        //dd($trick->getPicture());
-        $this->deleteFile->delete($trick->getPicture());
-        foreach ($img as $image){
-            $filename = "../public/uploads/pictures/";
-            if (file_exists($filename . $image->getName())) {
-                echo "Le fichier" . $image->getName() . "existe.";
-                //($image->getName());
-                unlink($filename . $image->getName());
-            } else {
-                echo "Le fichier" . $image->getName() .  "n'existe pas.";
+        $ok = $this->tokenStorage->getToken()->getUser();
+        $video = $this->getDoctrine()
+            ->getRepository(Video::class)
+            ->findOneBy(['trickId' => $id]);
+        $trick = $this->getDoctrine()
+            ->getRepository(Trick::class)
+            ->find($id);
+        $img = $this->getDoctrine()
+            ->getRepository(Image::class)
+            ->findBy(['trickId' => $id]);
+        if ($ok == $trick->getUser()) {
+            $this->deleteFile->delete($trick->getPicture());
+            foreach ($img as $image){
+                $filename = "../public/uploads/pictures/";
+                if (file_exists($filename . $image->getName())) {
+                    echo "Le fichier" . $image->getName() . "existe.";
+                    //($image->getName());
+                    unlink($filename . $image->getName());
+                } else {
+                    echo "Le fichier" . $image->getName() .  "n'existe pas.";
+                }
             }
-        }
-        
-        // ... perform some action, such as saving the task to the database
-        // for example, if Task is a Doctrine entity, save it!
-        
-        
-        //dd($img->getName());
-        //$fileSystem->exists('%kernel.project_dir%/public/uploads/pictures/');
-        //dd($fileSystem->exists(['public/uploads/pictures/']));
-        
-         $entityManager = $this->getDoctrine()->getManager();
-         $entityManager->Remove($video);
-         $entityManager->Remove($trick);
-         $entityManager->flush();
+            
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->Remove($video);
+            $entityManager->Remove($trick);
+            $entityManager->flush();
 
-         $this->session->getFlashBag()->add(
+            $this->session->getFlashBag()->add(
+                'success',
+                'Trick supprimé!'
+            );
+            
+            return $this->redirectToRoute('home');
+        }
+        $this->session->getFlashBag()->add(
             'success',
-            'Trick supprimé!'
+            'Vous n\'avez pas acces a cette page!'
         );
-        
         return $this->redirectToRoute('home');
-        
     }
 }
