+++ src/Controller/VideoDeleteController.php
@@ -2,11 +2,8 @@
 
 namespace App\Controller;
 
-use App\Entity\Trick;
+use App\Entity\Video;
 use App\Repository\VideoRepository;
-use Symfony\Component\HttpFoundation\Response;
-use App\Handlers\TrickAddHandler;
-use App\Responders\TrickAddResponder;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,49 +11,50 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class VideoDeleteController extends AbstractController
 {
-    /** @var TrickAddHandler */
-    //private $handler;
-    /** @var Responder */
-    private $responder;
-    /** @var EntityManagerInterface */
     private $session;
+    private $tokenStorage;
 
     public function __construct(
-        SessionInterface $session
+        SessionInterface $session,
+        TokenStorageInterface $tokenStorage
     
     ){
         $this->session = $session;
+        $this->tokenStorage = $tokenStorage;
     }
 
     /**
-    * @Route("/delete_video/{id}", name="delete_video")
+    * @Route("/delete_video/{id}/{trickid}", name="delete_video")
     * @IsGranted("ROLE_ADMIN")
     */
-    public function delete($id, VideoRepository $videoRepository)
+    public function delete($id, $trickid)
     {
-        // creates a task object and initializes some data for this example
-        //$task->setTask('Write a blog post');
-        //$task->setDueDate(new \DateTime('tomorrow'));
-        //var_dump($id);
-        
-        $video = $videoRepository
-        ->find($id);
+        $ok = $this->tokenStorage->getToken()->getUser();
+        $trick = $this->getDoctrine()
+            ->getRepository(Trick::class)
+            ->find($trickid);
+        $video = $this->getDoctrine()
+            ->getRepository(Video::class)
+            ->find($id);
+        if($ok == $trick->getUser()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->Remove($video);
+            $entityManager->flush();
 
-        // ... perform some action, such as saving the task to the database
-        // for example, if Task is a Doctrine entity, save it!
-        //dd($video);
-         $entityManager = $this->getDoctrine()->getManager();
-         $entityManager->Remove($video);
-         $entityManager->flush();
-
-         $this->session->getFlashBag()->add(
+            $this->session->getFlashBag()->add(
+                'success',
+                'Video supprimé!'
+            );
+            return $this->redirectToRoute('comment', ['id' => $trickid, 'page' => '1']);
+        }
+        $this->session->getFlashBag()->add(
             'success',
-            'Video supprimé!'
+            'Vous n\'avez pas acces a cette page!'
         );
-        
         return $this->redirectToRoute('home');
     }
 }
